// packages/raise/src/organization/dto/appearance.dto.ts
export class AppearancenDto {
  ownerUserId: String;
  ownerRealmId: String;
  ourStory: string;
  whyShouldWe: string;
  peopleSay: string;
  detailStory1: string;
  detailStory2: string;
  detailStory3: string;
  videoUrl: string;
  whySupportUs1: string;
  whySupportUs2: string;
  whySupportUs3: string;
  mainImageUrl: string;
  secondaryImage: string;
  eventImagesUrl1: string;
  eventImagesUrl2: string;
  eventImagesUrl3: string;
  favIcon: string;
  themesColor: string;
  usePallete: boolean;
  logo: string;
  headerAndFooter: string;
  accent: string;
  lButton: string;
}